+++ packages/commonwealth/client/scripts/views/pages/user_dashboard/user_dashboard_row.tsx
@@ -1,93 +1,25 @@
-import 'components/user_dashboard_row.scss';
+/* @jsx m */
 
+import m from 'mithril';
+import moment from 'moment';
 import { Icons, Button, MenuItem, PopoverMenu } from 'construct-ui';
 import { capitalize } from 'lodash';
-import m from 'mithril';
+
+import 'components/user_dashboard_row.scss';
+
 import {
   CWEvent,
   Label as ChainEventLabel,
   // CompoundEvents
 } from 'chain-events/src';
-import moment from 'moment';
-
 import app from 'state';
 import { NotificationCategories } from 'common-common/src/types';
-import {
-  AddressInfo,
-  DashboardActivityNotification,
-  NotificationSubscription,
-} from 'models';
-
-import { QuillFormattedText } from 'views/components/quill/quill_formatted_text';
-import { MarkdownFormattedText } from 'views/components/quill/markdown_formatted_text';
+import { AddressInfo, DashboardActivityNotification } from 'models';
 import User from 'views/components/widgets/user';
 import { formatTimestamp } from 'helpers/index';
-
-import { notifySuccess } from 'controllers/app/notifications';
-import { getProposalUrlPath } from '../../identifiers';
-import { CWCommunityAvatar } from './component_kit/cw_community_avatar';
-
-const getCommentPreview = (commentText) => {
-  // TODO Graham 6-5-22: Duplicate with notification_row.ts? See relevant note there
-  let decodedCommentText;
-  try {
-    const doc = JSON.parse(decodeURIComponent(commentText));
-    if (!doc.ops) throw new Error();
-    decodedCommentText = m(QuillFormattedText, {
-      doc,
-      hideFormatting: true,
-      collapse: false,
-    });
-  } catch (e) {
-    let doc = decodeURIComponent(commentText);
-    const regexp = RegExp('\\[(\\@.+?)\\]\\(.+?\\)', 'g');
-    const matches = doc['matchAll'](regexp);
-    Array.from(matches).forEach((match) => {
-      doc = doc.replace(match[0], match[1]);
-    });
-    decodedCommentText = m(MarkdownFormattedText, {
-      doc: doc.slice(0, 140),
-      hideFormatting: true,
-      collapse: true,
-    });
-  }
-  return decodedCommentText;
-};
-
-// Subscriptions
-const subscribeToThread = async (
-  threadId: string,
-  bothActive: boolean,
-  commentSubscription: NotificationSubscription,
-  reactionSubscription: NotificationSubscription
-) => {
-  const adjustedId = `discussion_${threadId}`;
-  if (bothActive) {
-    await app.user.notifications.disableSubscriptions([
-      commentSubscription,
-      reactionSubscription,
-    ]);
-    notifySuccess('Unsubscribed!');
-  } else if (!commentSubscription || !reactionSubscription) {
-    await Promise.all([
-      app.user.notifications.subscribe(
-        NotificationCategories.NewReaction,
-        adjustedId
-      ),
-      app.user.notifications.subscribe(
-        NotificationCategories.NewComment,
-        adjustedId
-      ),
-    ]);
-    notifySuccess('Subscribed!');
-  } else {
-    await app.user.notifications.enableSubscriptions([
-      commentSubscription,
-      reactionSubscription,
-    ]);
-    notifySuccess('Subscribed!');
-  }
-};
+import { getProposalUrlPath } from 'identifiers';
+import { CWCommunityAvatar } from '../../components/component_kit/cw_community_avatar';
+import { subscribeToThread, getCommentPreview } from './helpers';
 
 // Discuss, Share, and Subscribe Buttons
 const ButtonRow: m.Component<{
@@ -341,20 +273,20 @@ const ActivityContent: m.Component<{
   },
 };
 
-const UserDashboardRow: m.Component<
-  {
-    notification: DashboardActivityNotification;
-    onListPage?: boolean;
-  },
-  {
-    Labeler: any;
-    MolochTypes: any;
-    SubstrateTypes: any;
-    scrollOrStop: boolean;
-    markingRead: boolean;
-  }
-> = {
-  view: (vnode) => {
+export class UserDashboardRow
+  implements
+    m.ClassComponent<{
+      notification: DashboardActivityNotification;
+      onListPage?: boolean;
+    }>
+{
+  Labeler: any;
+  MolochTypes: any;
+  SubstrateTypes: any;
+  scrollOrStop: boolean;
+  markingRead: boolean;
+
+  view(vnode) {
     const {
       likeCount,
       viewCount,
@@ -467,7 +399,5 @@ const UserDashboardRow: m.Component<
         ]),
       ]
     );
-  },
-};
-
-export default UserDashboardRow;
+  }
+}
